od4.6 : Fixed warnings on Cdash ( Nightly Build )
Summary:
Warnings on cdash ( win64_od4.6_Release_Nightly info )
http://intranet/cdash/index.php?project=OpendTect-4.6
Thursday, June 20 2013 - 02:00 CEST

Test Plan: compiles

Reviewers: ranojay

Reviewed By: ranojay

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D1251

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@30383 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiFlatView/flatviewaxesdrawer.cc
@@ -104,8 +104,9 @@ void FlatView::AxesDrawer::setViewRect( const uiRect& rect )
     }
 
     if ( !rectitem_ )
-	rectitem_ =  view_.scene().addRect( rect.left(), rect.top(),
-					    rect.width(), rect.height() );
+	rectitem_ =  view_.scene().addRect( 
+				(float) rect.left(), (float) rect.top(),
+				(float) rect.width(), (float) rect.height() );
     else
 	rectitem_->setRect( rect.left(), rect.top(),
 			    rect.width(), rect.height() );
